+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Tests for `MaterialUtils`
 - Tests for `Colors`
 - Tests for `TransformUtils`
+- Tests for `ShapeCreator` (incomplete)
 
 ## [0.1.0] / 2023-10-01
 ### Quaternion